move preprocessing to utils
